requirements.txt
numpy
pandas
pydeck
requests
seaborn
streamlit
google-cloud-bigquery
google-cloud
google-cloud-core
vertexai
